+++ main.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"github.com/life2you/datas-go/handler"
 	"os"
 	"os/signal"
 	"syscall"
@@ -36,38 +37,19 @@ func main() {
 
 	// 5. 配置WebSocket
 	configs.GlobalConfig.WebSocket.OnConnect = rpcCallBack
-
 	// 如果RPC配置中有代理URL，则使用它
 	if configs.GlobalConfig.Proxy.Enabled && configs.GlobalConfig.Proxy.URL != "" {
 		logger.Info("使用代理连接Helius", zap.String("proxy", configs.GlobalConfig.Proxy.URL))
 		configs.GlobalConfig.WebSocket.ProxyURL = configs.GlobalConfig.Proxy.URL
 		configs.GlobalConfig.HeliusAPI.ProxyURL = configs.GlobalConfig.Proxy.URL
 		configs.GlobalConfig.HeliusEnhancedAPI.ProxyURL = configs.GlobalConfig.Proxy.URL
+		configs.GlobalConfig.PumpPortal.ProxyURL = configs.GlobalConfig.Proxy.URL
 	}
-
-	// 6. 初始化WebSocket客户端
-	rpc.NewWebSocketClientOptions(&configs.GlobalConfig.WebSocket)
-	if rpc.GlobalWebSocketClient == nil {
-		logger.Fatal("WebSocket客户端初始化失败")
-	}
-	logger.Info("WebSocket客户端初始化成功")
-
-	// 6.1 初始化Helius HTTP API客户端
-	rpc.NewHeliusClient(&configs.GlobalConfig.HeliusAPI)
-	if rpc.GlobalHeliusClient == nil {
-		logger.Fatal("Helius HTTP API客户端初始化失败")
-	}
-	logger.Info("Helius HTTP API客户端初始化成功")
-
-	// 6.2 初始化Helius Enhanced API客户端
-	rpc.NewHeliusEnhancedApiClient(&configs.GlobalConfig.HeliusEnhancedAPI)
-	if rpc.GlobalHeliusEnhancedApiClients == nil || len(rpc.GlobalHeliusEnhancedApiClients) == 0 {
-		logger.Fatal("Helius Enhanced API客户端初始化失败")
-	}
-	logger.Info("Helius Enhanced API客户端初始化成功")
-
+	rpc.NewPumpPortalClient(&configs.GlobalConfig.PumpPortal, handler.PumpPortalHandler)
+	service.StartPumpPortalService()
+	//initClient()
 	// 7. 启动服务，不需要阻塞
-	initStartService()
+	// initStartService()
 
 	// 8. 在主协程中打印状态信息
 	logger.Info("程序已启动，正在等待区块数据...")
@@ -92,12 +74,36 @@ func main() {
 	select {}
 }
 
+func initClient() {
+	// 6. 初始化WebSocket客户端
+	rpc.NewWebSocketClientOptions(&configs.GlobalConfig.WebSocket)
+	if rpc.GlobalWebSocketClient == nil {
+		logger.Fatal("WebSocket客户端初始化失败")
+	}
+	logger.Info("WebSocket客户端初始化成功")
+
+	// 6.1 初始化Helius HTTP API客户端
+	rpc.NewHeliusClient(&configs.GlobalConfig.HeliusAPI)
+	if rpc.GlobalHeliusClient == nil {
+		logger.Fatal("Helius HTTP API客户端初始化失败")
+	}
+	logger.Info("Helius HTTP API客户端初始化成功")
+
+	// 6.2 初始化Helius Enhanced API客户端
+	rpc.NewHeliusEnhancedApiClient(&configs.GlobalConfig.HeliusEnhancedAPI)
+	if rpc.GlobalHeliusEnhancedApiClients == nil || len(rpc.GlobalHeliusEnhancedApiClients) == 0 {
+		logger.Fatal("Helius Enhanced API客户端初始化失败")
+	}
+	logger.Info("Helius Enhanced API客户端初始化成功")
+}
+
 func initStartService() {
 	service.StartHeliusService()
 	time.Sleep(5 * time.Second)
 	service.ScanBlockQueue()
 	service.ProcessTransactionQueue()
-	logger.Info("所有服务已启动: 区块队列扫描服务、交易队列处理服务")
+	service.StartPumpPortalService()
+	logger.Info("所有服务已启动: 区块队列扫描服务、交易队列处理服务、PumpPortal服务")
 }
 
 func initQueue() {