+++ parser/transfer.go
@@ -1,4 +1,4 @@
-package transaction
+package parser
 
 import "github.com/life2you/datas-go/models/resp"